Fix for ModelsBuilderConfig
