+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -12,6 +12,7 @@
 
     <ItemGroup>
       <PackageReference Include="System.ComponentModel.Annotations" Version="4.6.0" />
+      <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.6.0" />
       <PackageReference Include="System.Reflection.Emit.Lightweight" Version="4.3.0" />
       <PackageReference Include="System.Runtime.Caching" Version="4.6.0" />
     </ItemGroup>
